Implementing smooth traversal over terrain by querying height of terrain at players x, z

+++ Source/Engine/EngineConfig.h
@@ -4,8 +4,8 @@ namespace Engine::EngineConfig {
 
 	class EngineConfig {
 	public:
-		static constexpr size_t MAX_ENTITIES = 3;
-		static constexpr size_t MAX_TEXTURES = 128;
-		static constexpr size_t MAX_MATERIALS = 256;
+		static constexpr uint16_t MAX_ENTITIES = 3;
+		static constexpr uint16_t MAX_TEXTURES = 128;
+		static constexpr uint16_t MAX_MATERIALS = 256;
 	};
 }
